Enable HSTS Preload
HSTS is the great little response header that tells a browser to always
use SSL/TLS to communicate with your site. It doesn't matter if the
user, or a link they are clicking, specifies HTTP, HSTS will remove the
ability for a compatible browser to use HTTP and will enforce the use of
HTTPS. This still leaves a user vulnerable on their very first
connection to a site though. HSTS preloading fixes that.